Added error handling

+++ src/options.rs
@@ -109,5 +109,5 @@ pub struct Options {
 
     // SOA Values
     #[clap(long, env = "SOA_VALUES", default_value = "1 60 60 31356000 0", num_args = 5, value_delimiter = ' ', default_value("1 60 60 86400 86400"), required = false)]
-    pub soa_values: Vec<u32>,
+    pub soa_values: Vec<i64>,
 }
